+++ package.json
@@ -6,7 +6,7 @@
     "build": "next build",
     "dev": "next dev",
     "postinstall": "prisma generate",
-    "lint": "next lint",
+    "lint": "next lint && prettier -c .",
     "start": "next start",
     "prepare": "husky install"
   },
